Add content endpoint route hooks and move content code to own file
